backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
slowapi>=0.1.9
sqlalchemy==2.0.36
python-dotenv==1.0.0
httpx==0.25.1
google-generativeai==0.8.5
python-multipart==0.0.6
facebook-business>=18.0.0
psycopg2-binary==2.9.11
fal-client
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
email-validator>=2.0.0
boto3>=1.34.0
alembic
